+++ tests/test_09_rp_service.py
@@ -21,7 +21,7 @@ from fedservice.entity.function import collect_trust_chains
 from fedservice.entity.function import verify_trust_chains
 from fedservice.fetch_entity_statement.fs2 import FSFetchEntityStatement
 from fedservice.fetch_entity_statement.fs2 import FSPublisher
-from fedservice.rp import ClientEntity
+from fedservice.appclient import ClientEntity
 from .utils import DummyCollector
 
 BASE_PATH = os.path.abspath(os.path.dirname(__file__))
