+++ tests/test_60_automatic_registration.py
@@ -12,7 +12,7 @@ from fedservice.defaults import LEAF_ENDPOINTS
 from fedservice.op import ServerEntity
 from fedservice.op.authorization import Authorization
 from fedservice.op.registration import Registration
-from fedservice.rp import ClientEntity
+from fedservice.appclient import ClientEntity
 from fedservice.utils import make_federation_combo
 from fedservice.utils import make_federation_entity
 from . import create_trust_chain_messages
@@ -98,7 +98,7 @@ class TestAutomatic(object):
 
         oidc_service = DEFAULT_OIDC_SERVICES.copy()
         oidc_service.update(DEFAULT_OIDC_FED_SERVICES)
-        oidc_service['authorization'] = {"class": "fedservice.rp.authorization.Authorization"}
+        oidc_service['authorization'] = {"class": "fedservice.appclient.oidc.authorization.Authorization"}
 
         self.rp = make_federation_combo(
             entity_id=RP_ID,
